+++ module/Admission/src/Admission/Assertion/AdmissionAssertion.php
@@ -269,7 +269,7 @@ class AdmissionAssertion extends AdmissionAbstractAssertion implements UserConte
                 $admission->getEtat()->getCode() == (Admission::ETAT_EN_COURS_VALIDATION || Admission::ETAT_VALIDE || Admission::ETAT_REJETE),
                 "Le dossier d'admission doit être en cours de validation / être validé / être rejeté"
             );
-        }elseif($codeNatureFichier === NatureFichier::CODE_ADMISSION_CHARTE_DOCTORAT_SIGNEE){
+        }elseif($codeNatureFichier === NatureFichier::CODE_ADMISSION_CHARTE_DOCTORAT_SIGNEE || $codeNatureFichier === NatureFichier::CODE_ADMISSION_CONVENTION_SIGNEE){
             if(!($this->userContextService->getSelectedIdentityRole()->getRoleId() === Role::ROLE_ID_ADMISSION_CANDIDAT || $this->userContextService->getSelectedRoleEcoleDoctorale() || $this->userContextService->getSelectedIdentityRole()->getCode() === Role::CODE_ADMIN_TECH)){
                 throw new FailedAssertionException("Seul le candidat et le gestionnaire d'ED peut gérer ce document");
             }